--- scr/database/models.py
@@ -1,62 +1,0 @@
-from sqlalchemy import Column, Integer, String, Boolean, func, Table
-from sqlalchemy.orm import relationship
-from sqlalchemy.sql.schema import ForeignKey
-from sqlalchemy.sql.sqltypes import DateTime
-from sqlalchemy.ext.declarative import declarative_base
-from typing import List
-
-Base = declarative_base()
-
-
-image_m2m_tag = Table(
-    "note_m2m_tag",
-    Base.metadata,
-    Column("id", Integer, primary_key=True),
-    Column("image_id", Integer, ForeignKey("images.id", ondelete="CASCADE")),
-    Column("tag_id", Integer, ForeignKey("tags.id", ondelete="CASCADE")),
-)
-
-
-class Image(Base):
-    __tablename__ = "images"
-    id = Column(Integer, primary_key=True)
-    title = Column(String(50), nullable=False)
-    url = Column(String(255), nullable=False)
-    created_at = Column('created_at', DateTime, default=func.now())
-    description = Column(String(150), nullable=False)
-    # comments = relationship(List('comments'))
-    tags = relationship("Tag", secondary=image_m2m_tag, backref="images")
-
-
-class Tag(Base):
-    __tablename__ = "tags"
-    id = Column(Integer, primary_key=True)
-    name_tag = Column(String(25), nullable=False, unique=True)
-
-
-class Comment(Base):
-    __tablename__ = "comments"
-    id = Column(Integer, primary_key=True)
-    comment = Column(String(255), nullable=False)
-
-
-class User(Base):
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True)
-    nickname = Column(String(30), nullable=False, unique=True)
-    email = Column(String(100), nullable=False, unique=True)
-    password = Column(String(255), nullable=False)
-    # images = relationship(List('images'))
-    avatar = relationship('images')
-    created_at = Column('created_at', DateTime, default=func.now())
-    refresh_token = Column(String(255))
-    # logout_token = Column(String(255))
-    confirmed = Column(Boolean, default=False)
-    banned = Column(Boolean, default=False)
-    role = relationship('users_roles')
-
-
-class Role(Base):
-    __tablename__ = 'users_roles'
-    id = Column(Integer, primary_key=True)
-    role_name = Column(String(30), nullable=False, unique=True)
